+++ eelbrain/_stats/stats.py
@@ -1,14 +1,16 @@
 # Author: Christian Brodbeck <christianbrodbeck@nyu.edu>
 """Statistics functions that work on numpy arrays."""
+from __future__ import annotations
+
 from dataclasses import dataclass
 import re
-from typing import Optional, Sequence, Union
+from typing import Literal, Sequence, Union
 
 import numpy as np
 import scipy.stats
 from scipy.linalg import inv
 
-from .._data_obj import Categorial, CellArg, Factor, Model, Parametrization, asfactor, asmodel
+from .._data_obj import CategorialArg, CellArg, Dataset, FactorArg, Model, Parametrization, asarray, ascategorial, asfactor, asmodel
 from . import opt
 from . import vector
 
@@ -19,7 +21,7 @@ FLOAT64 = np.dtype('float64')
 @dataclass
 class DispersionSpec:
     multiplier: float = 1
-    measure: str = 'SEM'  # Literal['SEM', 'CI']
+    measure: Literal['SEM', 'CI', 'SD'] = 'SEM'
 
     @classmethod
     def from_string(cls, string: Union[str, 'DispersionSpec']):
@@ -229,7 +231,7 @@ def rmssd(y):
     return np.sqrt(x)
 
 
-class SEM:
+class Dispersion:
 
     def __init__(self, y, x=None, match=None):
         """Standard error of the mean (SEM)
@@ -266,7 +268,7 @@ class SEM:
         self.model = model
         self.sem = sem
 
-    def get(self, spec: DispersionSpec):
+    def get(self, spec: Union[str, DispersionSpec]):
         spec = DispersionSpec.from_string(spec)
         if spec.measure == 'SEM':
             return self.sem * spec.multiplier
@@ -454,15 +456,16 @@ def ttest_t(p, df, tail=0):
     return t
 
 
-def variability(
+def dispersion(
         y: np.ndarray,
-        x: Optional[Categorial],
-        match: Factor,
-        spec: str,
-        pool: bool,
+        x: CategorialArg = None,
+        match: FactorArg = None,
+        spec: str = 'SEM',
         cells: Sequence[CellArg] = None,
+        data: Dataset = None,
+        pool: bool = None,
 ):
-    """Calculate data variability
+    """Calculate data dispersion measure
 
     Parameters
     ----------
@@ -481,7 +484,8 @@ def variability(
         ``99%ci``: 99% confidence interval.
     pool
         Pool the variability to create a single estimate (as opposed to one for
-        each cell in x).
+        each cell in x). Default ``True`` iff ``match`` is specified (for
+        within-subject errors).
     cells
         Estimate variance in these cells of ``x`` (default ``x.cells``).
 
@@ -492,11 +496,18 @@ def variability(
         an estimate for every cell in x.
     """
     spec_ = DispersionSpec.from_string(spec)
+    y, n = asarray(y, data=data, return_n=True)
+    if match is not None:
+        match = asfactor(match, data=data, n=n)
+    if pool is None:
+        pool = match is not None
     if x is None:
         if match is not None and match.df == len(match) - 1:
             raise ValueError("Can't calculate within-subject error because the match predictor explains all variability")
-    elif not pool and cells is None:
-        cells = x.cells
+    else:
+        x = ascategorial(x, data=data, n=n)
+        if not pool and cells is None:
+            cells = x.cells
 
     y = np.asarray(y, np.float64)
     if spec_.measure == 'SD':
@@ -506,11 +517,11 @@ def variability(
         if spec_.multiplier != 1:
             out *= spec_.multiplier
     elif pool or x is None:
-        out = SEM(y, x, match).get(spec_)
+        out = Dispersion(y, x, match).get(spec_)
     elif match is not None:
         raise NotImplementedError(f"{spec!r} unpooled with match")
     else:
-        out = np.array([SEM(y[x == cell]).get(spec_) for cell in cells])
+        out = np.array([Dispersion(y[x == cell]).get(spec_) for cell in cells])
 
     # return scalars for 1d-arrays
     if out.ndim == 0: