+++ doc/changes.rst
@@ -18,9 +18,18 @@ Major changes
 New in 0.40
 -----------
 
+* Due to a change in randomization algorithm, the permutation distribution for vector-based mass univariate tests may be different (old and new versions are both valid randomizations).
 * :func:`boosting` now accepts data with ragged trials (trials of different lengths).
 * :func:`boosting` now stores both the l1 and l2 loss of the final fit.
-* API: Plotting parameters ``ncol`` and ``nrow`` have been renamed to ``columns`` and ``rows``
+* Pipeline:
+
+    - :class:`pipeline.RawICA` preprocessing pipe now allows setting ``fit_kwargs``.
+
+* API:
+
+   - Plotting parameters ``ncol`` and ``nrow`` have been renamed to ``columns`` and ``rows``.
+   - :class:`Factor` cells that are not specified in ``labels`` are now ordered by their first ocurrence in ``x`` (previously order was alphabetic).
+   - The ``pool_error`` argument is replaced by ``within_subject_error``.
 
 
 New in 0.39